Move declarative test to its own class
